refactor: move actions to separate module
